refactor: remove backward-compat re-exports from LLM __init__
